Refactored project names and file namespaces
