--- ADF.Functions/host.json
@@ -1,18 +1,0 @@
-{
-    "version": "2.0",
-    "logging": {
-        "applicationInsights": {
-            "samplingExcludedTypes": "Request",
-            "samplingSettings": {
-                "isEnabled": true
-            }
-        },
-      "logLevel": {
-        "Azure.Datafactory.Extensions.Functions.DataLakeFunctions": "Information",
-        "Azure.Datafactory.Extensions.Functions.FunctionsBase": "Information",
-        "Azure.Datafactory.Extensions.Functions.DataLakeConfigFactory": "Information",
-        "Azure.Datafactory.Extensions.DataLake.DataLakeClientFactory": "Information",
-        "Azure.Datafactory.Extensions.DataLake.DataLakeControllerFactory": "Information"
-      }
-    }
-}
