+++ Azure.DataPipelineTools.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30225.117
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ADF.Functions", "ADF.Functions\ADF.Functions.csproj", "{12DD4B77-8B6C-4F4F-B0D2-58E60D2735BF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataPipelineTools.Functions", "DataPipelineTools.Functions\DataPipelineTools.Functions.csproj", "{12DD4B77-8B6C-4F4F-B0D2-58E60D2735BF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.DataFactory.Extensions", "Azure.DataFactory.Extensions\Azure.DataFactory.Extensions.csproj", "{84B7E7A9-BDE5-4602-B99C-853CC1159CE5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataPipelineTools", "DataPipelineTools\DataPipelineTools.csproj", "{84B7E7A9-BDE5-4602-B99C-853CC1159CE5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
