+++ DataPipelineTools.Functions/DataPipelineTools.Functions.csproj
@@ -11,11 +11,11 @@
     <Company />
     <Authors>Niall Langley</Authors>
     <Product>azure.datafactory.extensions.functions</Product>
-    <PackageId>Azure.Datafactory.Extensions.Functions</PackageId>
+    <PackageId>SqlCollaborative.Azure.DataPipelineTools.Functions</PackageId>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageProjectUrl>https://github.com/NJLangley/azure.datafactory.extensions</PackageProjectUrl>
-    <AssemblyName>Azure.Datafactory.Extensions.Functions</AssemblyName>
-    <RootNamespace>Azure.Datafactory.Extensions.Functions</RootNamespace>
+    <PackageProjectUrl>https://github.com/sqlcollaborative/AzureDataPipelineTools</PackageProjectUrl>
+    <AssemblyName>SqlCollaborative.Azure.DataPipelineTools.Functions</AssemblyName>
+    <RootNamespace>SqlCollaborative.Azure.DataPipelineTools.Functions</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Azure.Identity" Version="1.3.0" />
@@ -40,6 +40,6 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Azure.DataFactory.Extensions\Azure.DataFactory.Extensions.csproj" />
+    <ProjectReference Include="..\DataPipelineTools\DataPipelineTools.csproj" />
   </ItemGroup>
 </Project>
