+++ DataPipelineTools.Functions/Properties/launchSettings.json
