+++ DataPipelineTools.Functions/Properties/serviceDependencies.json
