+++ DataPipelineTools.Functions/Properties/serviceDependencies.local.json
