+++ DataPipelineTools/DataPipelineTools.csproj
@@ -10,12 +10,12 @@
     <Version>1.0.0</Version>
     <Company />
     <Authors>Niall Langley</Authors>
-    <Product>azure.datafactory.extensions</Product>
-    <PackageId>Azure.Datafactory.Extensions</PackageId>
+    <Product>SqlCollaborative.Azure.DataPipelineTools</Product>
+    <PackageId>SqlCollaborative.Azure.DataPipelineTools</PackageId>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <PackageProjectUrl>https://github.com/NJLangley/azure.datafactory.extensions</PackageProjectUrl>
-    <AssemblyName>Azure.Datafactory.Extensions</AssemblyName>
-    <RootNamespace>Azure.Datafactory.Extensions</RootNamespace>
+    <PackageProjectUrl>https://github.com/sqlcollaborative/AzureDataPipelineTools</PackageProjectUrl>
+    <AssemblyName>SqlCollaborative.Azure.DataPipelineTools</AssemblyName>
+    <RootNamespace>SqlCollaborative.Azure.DataPipelineTools</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
