+++ DataPipelineTools/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "Azure.DataFactory.Extensions": {
+    "SqlCollaborative.Azure.DataPipelineTools": {
       "commandName": "Project",
       "workingDirectory": "bin/Debug"
     }
